初步完成

+++ app/controllers/order_items_controller.rb
@@ -3,6 +3,11 @@ class OrderItemsController < ApplicationController
   before_action :set_order_item, only: [:destroy]
 
   def destroy
+    unless current_user.admin?
+      redirect_back(fallback_location: root_path, notice: "没有权限")
+      return
+    end
+
     if @order_item.destroy!
       redirect_back(fallback_location: root_path, notice: "订单项删除成功！")
     else
